Reuse ring render data for all ring systems
