+++ src/celengine/renderglsl.cpp
@@ -15,6 +15,7 @@
 #include <cmath>
 #include <cstddef>
 #include <memory>
+#include <optional>
 
 #include <celcompat/numbers.h>
 
@@ -115,94 +116,7 @@ void renderGeometryShadow_GLSL(Geometry* geometry,
     shadowFbo->unbind(oldFboId);
 }
 
-class GLRingRenderData : public RingRenderData
-{
-    static constexpr int nLODs = 4;
-public:
-    ~GLRingRenderData() override = default;
-
-    int count() const { return nLODs; }
-    bool isInitializedLOD(int i) const { return init[i]; };
-    void initializeLOD(int i, float innerRadius, float outerRadius, unsigned nSections);
-    void renderLOD(int i);
-
-private:
-    std::array<gl::Buffer, nLODs> bo;
-    std::array<gl::VertexObject, nLODs> vo;
-    std::array<bool, nLODs> init;
-};
-
-struct RingVertex
-{
-    std::array<float, 3>  pos;
-    std::array<unsigned short, 2> tex;
-};
-
-void
-GLRingRenderData::initializeLOD(int i, float innerRadius, float outerRadius, unsigned nSections)
-{
-    std::vector<RingVertex> ringCoord;
-    ringCoord.reserve(2 * nSections);
-
-    constexpr float angle = 2.0f * celestia::numbers::pi_v<float>;
-    for (unsigned i = 0; i <= nSections; i++)
-    {
-        float theta = angle * i / nSections; // NOSONAR
-        float s, c;
-        math::sincos(theta, s, c);
-
-        RingVertex vertex;
-        // inner point
-        vertex.pos[0] = c * innerRadius;
-        vertex.pos[1] = 0.0f;
-        vertex.pos[2] = s * innerRadius;
-        vertex.tex[0] = 0;
-        vertex.tex[1] = 0;
-        ringCoord.push_back(vertex);
-
-        // outer point
-        vertex.pos[0] = c * outerRadius;
-        // vertex.pos[1] = 0.0f;
-        vertex.pos[2] = s * outerRadius;
-        vertex.tex[0] = 1;
-        // vertex.tex[1] = 0;
-        ringCoord.push_back(vertex);
-    }
-
-    bo[i] = gl::Buffer(gl::Buffer::TargetHint::Array, ringCoord);
-    vo[i] = gl::VertexObject(gl::VertexObject::Primitive::TriangleStrip);
-    vo[i]
-        .setCount((nSections + 1) * 2)
-        .addVertexBuffer(
-            bo[i],
-            CelestiaGLProgram::TextureCoord0AttributeIndex,
-            2,
-            gl::VertexObject::DataType::UnsignedShort,
-            false,
-            sizeof(RingVertex),
-            offsetof(RingVertex, tex))
-        .addVertexBuffer(
-            bo[i],
-            CelestiaGLProgram::VertexCoordAttributeIndex,
-            3,
-            gl::VertexObject::DataType::Float,
-            false,
-            sizeof(RingVertex),
-            offsetof(RingVertex, pos));
-    init[i] = true;
-    bo[i].unbind();
-}
-
-void
-GLRingRenderData::renderLOD(int i)
-{
-    glDisable(GL_CULL_FACE);
-    vo[i].draw();
-    glEnable(GL_CULL_FACE);
-}
-
 } // end unnamed namespace
-
 
 // Render a planet sphere with GLSL shaders
 void renderEllipsoid_GLSL(const RenderInfo& ri,
@@ -723,141 +637,4 @@ void renderClouds_GLSL(const RenderInfo& ri,
                         textures.data(), static_cast<int>(textures.size()), prog);
 
     prog->textureOffset = 0.0f;
-}
-
-// Render a planetary ring system
-void renderRings_GLSL(RingSystem& rings,
-                      RenderInfo& ri,
-                      const LightingState& ls,
-                      float planetRadius,
-                      float planetOblateness,
-                      unsigned int textureResolution,
-                      bool renderShadow,
-                      float segmentSizeInPixels,
-                      const Matrices &m,
-                      bool inside,
-                      Renderer* renderer)
-{
-    float inner = rings.innerRadius / planetRadius;
-    float outer = rings.outerRadius / planetRadius;
-    Texture* ringsTex = rings.texture.find(textureResolution);
-
-    ShaderProperties shadprop;
-    // Set up the shader properties for ring rendering
-    {
-        shadprop.lightModel = LightingModel::RingIllumModel;
-        shadprop.nLights = std::min(ls.nLights, MaxShaderLights);
-
-        if (renderShadow)
-        {
-            // Set one shadow (the planet's) per light
-            for (unsigned int li = 0; li < ls.nLights; li++)
-                shadprop.setEclipseShadowCountForLight(li, 1);
-        }
-
-        if (ringsTex != nullptr)
-            shadprop.texUsage = TexUsage::DiffuseTexture;
-    }
-
-
-    // Get a shader for the current rendering configuration
-    auto* prog = renderer->getShaderManager().getShader(shadprop);
-    if (prog == nullptr)
-        return;
-
-    prog->use();
-    prog->setMVPMatrices(*m.projection, *m.modelview);
-
-    prog->eyePosition = ls.eyePos_obj;
-    prog->ambientColor = ri.ambientColor.toVector3();
-    prog->setLightParameters(ls, ri.color, ri.specularColor, Color::Black);
-
-    for (unsigned int li = 0; li < ls.nLights; li++)
-    {
-        const DirectionalLight& light = ls.lights[li];
-
-        // Compute the projection vectors based on the sun direction.
-        // I'm being a little careless here--if the sun direction lies
-        // along the y-axis, this will fail.  It's unlikely that a
-        // planet would ever orbit underneath its sun (an orbital
-        // inclination of 90 degrees), but this should be made
-        // more robust anyway.
-        Eigen::Vector3f axis = Eigen::Vector3f::UnitY().cross(light.direction_obj);
-        float cosAngle = Eigen::Vector3f::UnitY().dot(light.direction_obj);
-        axis.normalize();
-
-        float tScale = 1.0f;
-        if (planetOblateness != 0.0f)
-        {
-            // For oblate planets, the size of the shadow volume will vary
-            // based on the light direction.
-
-            // A vertical slice of the planet is an ellipse
-            float a = 1.0f;                          // semimajor axis
-            float b = a * (1.0f - planetOblateness); // semiminor axis
-            float ecc2 = 1.0f - (b * b) / (a * a);   // square of eccentricity
-
-            // Calculate the radius of the ellipse at the incident angle of the
-            // light on the ring plane + 90 degrees.
-            float r = a * std::sqrt((1.0f - ecc2) /
-                                    (1.0f - ecc2 * math::square(cosAngle)));
-
-            tScale *= a / r;
-        }
-
-        // The s axis is perpendicular to the shadow axis in the plane of the
-        // of the rings, and the t axis completes the orthonormal basis.
-        Eigen::Vector3f sAxis = axis * 0.5f;
-        Eigen::Vector3f tAxis = (axis.cross(light.direction_obj)) * 0.5f * tScale;
-        Eigen::Vector4f texGenS;
-        texGenS.head(3) = sAxis;
-        texGenS[3] = 0.5f;
-        Eigen::Vector4f texGenT;
-        texGenT.head(3) = tAxis;
-        texGenT[3] = 0.5f;
-
-        // r0 and r1 determine the size of the planet's shadow and penumbra
-        // on the rings.
-        // TODO: A more accurate ring shadow calculation would set r1 / r0
-        // to the ratio of the apparent sizes of the planet and sun as seen
-        // from the rings. Even more realism could be attained by letting
-        // this ratio vary across the rings, though it may not make enough
-        // of a visual difference to be worth the extra effort.
-        float r0 = 0.24f;
-        float r1 = 0.25f;
-        float bias = 1.0f / (1.0f - r1 / r0);
-
-        prog->shadows[li][0].texGenS = texGenS;
-        prog->shadows[li][0].texGenT = texGenT;
-        prog->shadows[li][0].maxDepth = 1.0f;
-        prog->shadows[li][0].falloff = bias / r0;
-    }
-
-    if (ringsTex != nullptr)
-        ringsTex->bind();
-
-    if (rings.renderData == nullptr)
-        rings.renderData = std::make_unique<GLRingRenderData>();
-    auto data = static_cast<GLRingRenderData*>(rings.renderData.get());
-
-    unsigned nSections = 180;
-    int i = 0;
-    for (i = 0; i < data->count() - 1; i++)
-    {
-        float s = segmentSizeInPixels * tan(celestia::numbers::pi / nSections);
-        if (s < 30.0f) // TODO: make configurable
-            break;
-        nSections <<= 1;
-    }
-
-    Renderer::PipelineState ps;
-    ps.blending = true;
-    ps.blendFunc = {GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA};
-    ps.depthTest = true;
-    ps.depthMask = inside;
-    renderer->setPipelineState(ps);
-
-    if (!data->isInitializedLOD(i))
-        data->initializeLOD(i, inner, outer, nSections);
-    data->renderLOD(i);
 }
