+++ src/celengine/renderglsl.h
@@ -19,14 +19,12 @@
 
 #include <celutil/reshandle.h>
 
-
 class Atmosphere;
 class Geometry;
 class LightingState;
 struct Matrices;
 class Renderer;
 struct RenderInfo;
-class RingSystem;
 class Texture;
 
 namespace celestia::math
@@ -71,18 +69,6 @@ void renderClouds_GLSL(const RenderInfo& ri,
                        const celestia::math::Frustum& frustum,
                        const Matrices &m,
                        Renderer* renderer);
-
-void renderRings_GLSL(RingSystem& rings,
-                      RenderInfo& ri,
-                      const LightingState& ls,
-                      float planetRadius,
-                      float planetOblateness,
-                      unsigned int textureResolution,
-                      bool renderShadow,
-                      float segmentSizeInPixels,
-                      const Matrices &m,
-                      bool inside,
-                      Renderer* renderer);
 
 void renderGeometry_GLSL_Unlit(Geometry* geometry,
                                const RenderInfo& ri,
