+++ src/celrender/rendererfwd.h
@@ -13,4 +13,5 @@ class LargeStarRenderer;
 class LineRenderer;
 class NebulaRenderer;
 class OpenClusterRenderer;
+class RingRenderer;
 }
